Add better message in case file exists

+++ dotfiles/create_links.py
@@ -6,12 +6,14 @@ dir_abspath = os.path.dirname(os.path.realpath(__file__))
 
 for f in os.listdir(dir_abspath):
     if f != os.path.basename(__file__):
-        source = "~/." + f
+        source = os.path.expanduser("~/." + f)
         target = os.path.join(dir_abspath, f)
 
-        print("Creating link from %s to %s" % (source, target))
-
-        cmd = "ln -s %s %s" % (target, source)
-        os.system(cmd)
+        if not os.path.exists(source):
+            print("Create: %s --> %s" % (source, target))
+            cmd = "ln -s %s %s" % (target, source)
+            os.system(cmd)
+        else:
+            print("%s exists already. Not overwriting." % source)
 
 print("Done!")
